move HIP 904 tests back
