+++ hedera-node/test-clients/src/main/java/com/hedera/services/bdd/suites/hip904/UnlimitedAutoAssociationSuite.java
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package com.hedera.services.bdd.suites.contract.hips.hip904;
+package com.hedera.services.bdd.suites.hip904;
 
 import static com.google.protobuf.ByteString.copyFromUtf8;
 import static com.hedera.node.app.hapi.utils.EthSigsUtils.recoverAddressFromPubKey;
